+++ examples/08_middleware.rs
@@ -1,7 +1,8 @@
 use afire::{Header, Method, Response, Server};
 
-// Read request headers and send a response with custom headers
-// You may want to read https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers to understand headers more
+// Define middleware that will be executed before the request handler.
+// In afire middleware are run before the normal routes and they can ether
+// send a response or pass the request to the next middleware / route.
 
 fn main() {
     // Create a new Server instance on localhost port 8080
@@ -49,9 +50,7 @@ fn main() {
 
     // Define server wide default headers
     // These will be send with every response
-    // If the same header is defined in the route it will be put before the default header
-    // Although it is not garunteed to be the one picked by the client it usually is
-    // At the bottom of this file is a representation of the order of the headers
+    // Default headers also have a higher priority than route specific headers so they will override them
     server.add_default_header(Header::new(
         "X-Server-Header",
         "This is a server wide header",
@@ -61,7 +60,7 @@ fn main() {
     // And you can goto http://localhost:8080/headers to see the headers your client sent to the server
 
     println!(
-        "[05] Listening on http://{}:{}",
+        "[08] Listening on http://{}:{}",
         server.ip_string(),
         server.port
     );
@@ -70,14 +69,3 @@ fn main() {
     // This will block the current thread
     server.start();
 }
-
-// This is a representation of the order of the headers
-// The order is important because the client usually picks the first header it can find
-
-// -------------------
-// |  Route Headers  |
-// -------------------
-// | Default Headers |
-// -------------------
-// |  ContentLength  |
-// -------------------